maalem-backend/requirements.txt
Django>=5.0
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.3.0
django-allauth>=0.58.0
django-cors-headers>=4.3.0
django-filter>=23.3
django-notifications-hq>=1.8.0
django-storages>=1.14
django-cloudinary-storage    
psycopg2-binary>=2.9.9
dj-database-url
gunicorn
daphne>=4.0.0
whitenoise
channels>=4.0.0
channels-redis>=4.1.0
python-dotenv>=1.0.0
Pillow>=10.0.0
celery>=5.3.4
redis>=5.0.1
boto3>=1.28
